Created a dir for Form and separated logic for product characteristics
