+++ client/src/components/RatingsReviews/Form/prodChars.jsx
@@ -1,0 +1,8 @@
+import React, { useState, useEffect } from 'react';
+
+const prodChars = () => {
+
+}
+
+
+export default prodChars;
