+++ client/src/components/RatingsReviews/RatingsReviews.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import ReviewList from './ReviewList.jsx';
 import { someReviews, noReviews } from './ExampleReviews.js';
 import please from '../../request.js';
-import Form from './Form.jsx';
+import Form from './Form/Form.jsx';
 
 const RatingsReviews = ({ productId, productName }) => {
